Moving impl to a different package
